docker/tools/.config/validation/core/testing/harness.py
#!/usr/bin/env python3
"""
Test Harness
Automated testing framework for validation rules.
"""
from __future__ import annotations

import logging
from pathlib import Path
from types import ModuleType
from typing import Any

_yaml: ModuleType | None
try:
    import yaml as _yaml_module

    _yaml = _yaml_module
except ImportError:
    _yaml = None

logger = logging.getLogger(__name__)


class TestHarness:
    """
    Rule testing harness.

    Executes rule tests and validates expected outcomes.
    """

    def __init__(self, rules_root: Path, test_data_root: Path):
        """
        Initialize test harness.

        Args:
            rules_root: Root directory of rule packages
            test_data_root: Root directory of test data
        """
        if not _yaml:
            raise ImportError("PyYAML required for test harness")

        self.rules_root = rules_root
        self.test_data_root = test_data_root

    def discover_tests(self, package: str | None = None) -> list[Path]:
        """
        Discover test files.

        Args:
            package: Package name (None = all packages)

        Returns:
            List of test file paths
        """
        if package:
            # Search specific package
            package_dir = self.rules_root / package / "tests"
            if not package_dir.exists():
                return []
            test_files = list(package_dir.glob("*.test.yaml"))
        else:
            # Search all packages
            test_files = list(self.rules_root.glob("*/tests/*.test.yaml"))

        logger.info("Discovered %s test files", len(test_files))
        return sorted(test_files)

    def load_test(self, test_file: Path) -> dict[str, Any]:
        """
        Load test definition.

        Args:
            test_file: Path to test file

        Returns:
            Test definition
        """
        with open(test_file, "r", encoding="utf-8") as f:
            test_def = _yaml.safe_load(f) if _yaml else None

        return test_def or {}

    def run_test(self, test_file: Path) -> dict[str, Any]:
        """
        Run single test.

        Args:
            test_file: Path to test file

        Returns:
            Test result
        """
        test_def = self.load_test(test_file)

        # Extract test metadata
        test_name = test_def.get("name", test_file.stem)
        rule_id = test_def.get("ruleID")
        test_cases = test_def.get("cases", [])

        results = []
        passed = 0
        failed = 0

        for case in test_cases:
            case_name = case.get("name", "unnamed")
            fixture = case.get("fixture")
            expect_match = case.get("expectMatch", True)

            # Run test case
            result = self._run_test_case(rule_id or "", fixture, expect_match)
            results.append(
                {
                    "name": case_name,
                    "passed": result["passed"],
                    "message": result.get("message", ""),
                }
            )

            if result["passed"]:
                passed += 1
            else:
                failed += 1

        return {
            "test_name": test_name,
            "rule_id": rule_id,
            "total": len(test_cases),
            "passed": passed,
            "failed": failed,
            "results": results,
        }

    def _run_test_case(self, rule_id: str, fixture: str, expect_match: bool) -> dict[str, Any]:  # pylint: disable=unused-argument
        """
        Run single test case.

        Args:
            rule_id: Rule identifier
            fixture: Fixture content or path
            expect_match: Whether rule should match

        Returns:
            Test case result
        """
        # This is a placeholder - actual implementation would:
        # 1. Apply rule to fixture
        # 2. Check if rule matched
        # 3. Compare with expectation

        # For now, return mock result
        return {
            "passed": True,
            "message": f"Test case for {rule_id} passed (mock)",
        }

    def run_all_tests(self, package: str | None = None) -> dict[str, Any]:
        """
        Run all tests.

        Args:
            package: Package name (None = all)

        Returns:
            Test suite results
        """
        test_files = self.discover_tests(package)

        results = []
        total_passed = 0
        total_failed = 0

        for test_file in test_files:
            try:
                result = self.run_test(test_file)
                results.append(result)
                total_passed += result["passed"]
                total_failed += result["failed"]
            except (OSError, ValueError, KeyError) as e:
                logger.error("Test failed: %s: %s", test_file.name, e)
                results.append(
                    {
                        "test_name": test_file.stem,
                        "error": str(e),
                    }
                )

        return {
            "total_tests": len(test_files),
            "total_cases": total_passed + total_failed,
            "passed": total_passed,
            "failed": total_failed,
            "results": results,
        }
