docker/tools/.config/validation/requirements.txt
# Python Requirements for Validation Framework
# Pinned versions for reproducibility (last updated: 2025-11-25)

# Core dependencies
pyyaml==6.0.2                    # YAML parsing and serialization
jsonschema==4.23.0               # JSON Schema validation
click==8.1.7                     # CLI framework

# Testing
pytest==8.3.3                    # Testing framework
pytest-cov==5.0.0                # Coverage reporting

# Code quality
black==24.10.0                   # Code formatting
ruff==0.7.4                      # Fast linting
mypy==1.13.0                     # Type checking

# Utilities
rich==13.9.4                     # Terminal formatting
tabulate==0.9.0                  # Table formatting
jinja2==3.1.4                    # Template engine (scaffolding)

# Optional: Documentation
mkdocs==1.6.1                    # Documentation generation
mkdocs-material==9.5.47          # Material theme
